sync: Make sure secrets are cleared only after they are no longer needed
This prevents situations when the session token is needed to obtain a
signed certificate + storage credentials to unregister the device at
sign out, but the session token had already been deleted. More specifically,
this is the case when the async requests take too long and the session
token ends up being deleted by _forget_tokens() before the requests complete.